restored game files (kept the new model files)
